requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# NLP and ML Libraries
nltk>=3.8.0
spacy>=3.5.0
spacy-curated-transformers>=0.3.1  # needed for en_core_web_trf pipeline
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0
camelot-py[cv]>=0.11.0
PyMuPDF>=1.22.0

# Web Framework
streamlit>=1.25.0
streamlit-js-eval>=0.1.5

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
regex>=2023.0.0
tqdm>=4.65.0
rapidfuzz>=3.0.0
keybert>=0.7.0
rake-nltk>=1.0.6
onnxruntime>=1.16.0
xgboost>=1.7.0
shap>=0.43.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
colorlog>=6.7.0
